+++ cypress/integration/BuyItem.spec.js
@@ -1,0 +1,16 @@
+import { setMockState } from '../fixtures/default'
+
+describe('Buy Item', () => {
+  beforeEach(() => {
+    setMockState()
+    cy.visit('http://localhost:3000')
+  })
+
+  it('Should buy an item.', () => {
+    cy.get('body').contains('100')
+    cy.contains('Market').click()
+    cy.get('#quantity-input-0').type('1')
+    cy.get('button[data-testid="add-button-0').click()
+    cy.get('body').contains('94')
+  })
+})