+++ src/App.js
@@ -1,53 +1,56 @@
 import React, { useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import { generatePlanets } from './util'
-import { setShipLocationValue, setShipLocationName } from './redux/actions/ship'
-import { setPlanets } from './redux/actions/world'
+import {
+  initializeApplication,
+  calculateContractExpirations
+} from './redux/actions/world'
 import View from './views/View'
 import { Box } from 'grommet'
-import CashDisplay from './components/CashDisplay'
-import ItemTimer from './components/ItemTimer'
-import Title from './components/Title'
-import ViewSelector from './components/ViewSelector'
+import Sidebar from './components/Sidebar'
 
-const App = ({ handleInitializeApplication, planets }) => {
+/**
+ * Hermes app.
+ */
+const App = ({
+  contracts,
+  handleContractExpirations,
+  handleInitializeApplication,
+  planets
+}) => {
   useEffect(() => {
     if (planets.length === 0) handleInitializeApplication()
+    if (contracts.length > 0) handleContractExpirations(contracts)
     // eslint-disable-next-line
-  }, [])
+  }, [contracts.length])
 
   return (
-    <Box fill>
-      <Title />
-      <ItemTimer />
-      <CashDisplay />
-      <ViewSelector />
-      <View />
+    <Box id="outer-container" fill>
+      <Sidebar outerContainerId="outer-container" pageWrapId="page-wrap" />
+      <Box id="page-wrap" margin={{ left: 'xlarge' }}>
+        <View />
+      </Box>
     </Box>
   )
 }
 
 App.propTypes = {
+  contracts: PropTypes.array.isRequired,
+  handleContractExpirations: PropTypes.func.isRequired,
   handleInitializeApplication: PropTypes.func.isRequired,
   planets: PropTypes.array.isRequired
 }
 
 const mapStateToProps = ({ world }) => ({
+  contracts: world.contracts,
   planets: world.planets
 })
 
 const mapDispatchToProps = dispatch => ({
-  handleInitializeApplication: () => {
-    const planets = generatePlanets()
-    const homePlanet = planets.find(planet => planet.isHomePlanet === true)
-    const value = homePlanet.location
-    const name = homePlanet.name
-
-    dispatch(setPlanets(planets))
-    dispatch(setShipLocationValue(value))
-    dispatch(setShipLocationName(name))
-  }
+  handleContractExpirations: contracts => {
+    if (!window.Cypress) dispatch(calculateContractExpirations(contracts))
+  },
+  handleInitializeApplication: () => dispatch(initializeApplication())
 })
 
 export default connect(