+++ src/components/PlanetDisplay.js
@@ -1,46 +1,43 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import { Box, Button, Heading } from 'grommet'
+import { Box, Button, Heading, Text } from 'grommet'
 import { Target } from 'grommet-icons'
-import {
-  setShipLocationName,
-  setShipLocationValue,
-  removeCargo
-} from '../redux/actions/ship'
-import { addCash } from '../redux/actions/user'
-import ItemDisplay from './ItemDisplay'
+import { departShip } from '../redux/actions/ship'
+import { createETA, createDiffDuration } from '../util'
 
-const PlanetDisplay = ({
-  handleShipTravel,
-  planet,
-  shipCargo,
-  shipLocationValue
-}) => {
-  const { isHomePlanet, items, location, name } = planet
+/**
+ * Displays planet statistics.
+ */
+const PlanetDisplay = ({ handleShipTravel, planet, ship }) => {
+  const { isHomePlanet, location, name } = planet
+  const shipLocationValue = ship.location.value
+
+  const eta = createETA({ value: planet.location }, ship)
+  const diffDuration = createDiffDuration(eta)
 
   return (
     <div key={name}>
-      <Box direction="row" gap="medium">
+      <Box align="center" direction="row" gap="medium">
         <Heading level="2">
           {isHomePlanet ? name + ' - Home Planet' : name}
         </Heading>
-        {shipLocationValue !== location && (
+        {shipLocationValue !== location && !ship.isShipTraveling && (
           <Button
             data-testid={`travel-button-${name}`}
             hoverIndicator
             icon={<Target />}
-            onClick={() =>
-              handleShipTravel({ name, value: location }, shipCargo)
-            }
+            onClick={() => handleShipTravel({ name, value: location }, ship)}
             plain
           />
         )}
+        {shipLocationValue !== location && !ship.isShipTraveling && (
+          <Text>
+            ETA: {diffDuration.minutes()} minutes {diffDuration.seconds()}{' '}
+            seconds
+          </Text>
+        )}
       </Box>
-      <span>Items:</span>
-      {items.map(item => (
-        <ItemDisplay key={item.id} item={item} location={location} />
-      ))}
     </div>
   )
 }
@@ -48,39 +45,14 @@ const PlanetDisplay = ({
 PlanetDisplay.propTypes = {
   handleShipTravel: PropTypes.func.isRequired,
   planet: PropTypes.object.isRequired,
-  shipCargo: PropTypes.object.isRequired,
-  shipLocationValue: PropTypes.number.isRequired
+  ship: PropTypes.object.isRequired
 }
 
-const mapStateToProps = ({ ship }) => ({
-  shipCargo: ship.cargo,
-  shipLocationValue: ship.location.value
-})
+const mapStateToProps = ({ ship }) => ({ ship })
 
 const mapDispatchToProps = dispatch => ({
-  handleShipTravel: (destination, shipCargo) => {
-    const sellableItems = shipCargo.items.filter(
-      item => item.destination.value === destination.value
-    )
-
-    let profit = 0
-
-    sellableItems.forEach(item => {
-      const { quantity, value } = item
-      const itemProfit = quantity * value
-      profit += itemProfit
-    })
-
-    dispatch(addCash(profit))
-
-    sellableItems.forEach(item => {
-      // * Remove the item from the ship cargo
-      dispatch(removeCargo(item))
-    })
-
-    dispatch(setShipLocationName(destination.name))
-    dispatch(setShipLocationValue(destination.value))
-  }
+  handleShipTravel: (destination, ship) =>
+    dispatch(departShip(destination, ship))
 })
 
 export default connect(