+++ src/redux/reducers/ship.js
@@ -1,8 +1,10 @@
 const shipDefaultState = {
   cargo: {
     items: [],
-    volumeRemaining: 5
+    volumeRemaining: 10 // m³
   },
+  destination: null,
+  isShipTraveling: false,
   location: {
     name: null,
     value: null
@@ -23,16 +25,24 @@ export default (state = shipDefaultState, action) => {
             state.cargo.volumeRemaining + action.payload.item.quantity
         }
       }
-    case 'SET_SHIP_LOCATION_NAME':
+    case 'REPLACE_SHIP':
+      return { ...action.payload.ship }
+    case 'SET_DESTINATION':
+      return { ...state, destination: action.payload.destination }
+    case 'SET_ETA':
       return {
         ...state,
-        location: { ...state.location, name: action.payload.name }
+        destination: { ...state.destination, eta: action.payload.eta }
       }
-    case 'SET_SHIP_LOCATION_VALUE':
+    case 'SET_SHIP_LOCATION':
       return {
         ...state,
-        location: { ...state.location, value: action.payload.value }
+        location: action.payload.location
       }
+    case 'SET_SHIP_TRAVELING':
+      return { ...state, isShipTraveling: action.payload.isShipTraveling }
+    case 'SET_TRAVEL_DURATION':
+      return { ...state, travelDuration: action.payload.travelDuration }
     case 'STORE_CARGO':
       let updatedItems = state.cargo.items.map(currentItem => {
         if (action.payload.item.id === currentItem.id) {