+++ CHANGELOG.md
@@ -1,16 +1,76 @@
 # Changelog
 
+## 📦 v0.4.0 (September 6, 2019)
+
+### Added
+
+- 💾 - Ability to import game from JSON file
+- 💾 - Ability to export game to JSON file
+- 📈 - Google Analytics
+- 🐛 - LogRocket
+- ✨ - Market view
+- ✨ - Ability to view item contracts
+- ✅ - Cypress Tests (Buy Item, Sell Item, Travel)
+
+### Changed
+
+- ✏️ - Use of NODE_ENV to call outside resources dynamically
+- ⏲️ - Item contracts expire at midnight of the day that they are created
+- 🚫 - No longer record video in Cypress tests if in Travis
+
+### Fixed
+
+- 🚫 - Disable Item button if the ship is traveling
+
+## 📦 v0.3.0 (August 30, 2019)
+
+### Added
+
+- ✨ - Ability to buy items
+- ✨ - Netlify Site
+- ✨ - ItemTimer has a heading to display time
+- ✨ - 7 new Item Types
+- ✨ - Item Contracts
+- ✨ - Travel Timer / ETA
+- ✨ - Cypress Integration
+- ✨ - Sidebar
+
+### Changed
+
+- ✏️ - Planets are now generated with unique ID's
+- ✏️ - Merged 2 actions into setShipLocation()
+- ✏️ - Item volume is more realistic
+- ✏️ - Item quantity is random
+- ✏️ - Adding an item costs the user cash
+- ✏️ - Items have a description
+
+### Fixed
+
+## 📦 v0.2.0 (August 23, 2019)
+
+### Added
+
+- ✨ - Item Count
+- ✨ - Testing Library
+
+### Changed
+
+- ✏️ - Item Timer based on real time
+- ✏️ - Additional badges in README
+
+### Fixed
+
 ## 📦 v0.1.1 (August 21, 2019)
 
 ### Added
 
-* 📝 - Changelog
-* ✨ - CashDisplay component
+- 📝 - Changelog
+- ✨ - CashDisplay component
 
 ### Changed
 
-* ✏️ - Refactored App.js
-* ✏️ - Refactored Redux
+- ✏️ - Refactored App.js
+- ✏️ - Refactored Redux
 
 ### Fixed
 
@@ -18,9 +78,8 @@
 
 ### Added
 
-* 💰 - Basic economy functionality and gameplay
+- 💰 - Basic economy functionality and gameplay
 
 ### Changed
 
 ### Fixed
-